requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2
mlflow==2.9.2
pyyaml==6.0.1
pytest==7.4.3
flake8==6.1.0
requests==2.31.0